Declare orientdb-core dependency to be of jar type
orientdb-core pom dependency is of no use here, it's not even a BOM pom. It is not only not-useful, but it is casusing issues. E.g. with SBT, even if I explicitly added orientdb-core dependency, the pom type of orientdb-core dependency in orientdb-graph pom was removing orientdb-core from the classpath, so in my code any class referencing a class which makes use of orientdb-core classes would not compile (was getting infamous "class file is broken error").

Thus, instead of pom type, orientdb-core dependency should be declared as dependency of jar type (default).