Add Rexster configuration class.
Rexster (http://rexster.tinkerpop.com)  no longer depends on OrientDB, therefore this class needs to live with the OrientDB repository. The addition of rexster-core to the pom comes with provided scope as OrientDB does not need to package rexster dependencies (they are provided by rexster when OrientDB is deployed there).