Fix #2787 - server.sh Java path spaces
Tested on: OSX 10.9.4
Added quotes around the path to make it possible to have a JAVA path (as defined in $JAVA) to work with spaces.
This happens when you have one java install at an Apple computer with a recent OSX version. In many tutorials about updating the java version this strategy is being used.